+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `SpaceLaunchTests` target and set it up for unit-testing.
 - Directory structure for the `SpaceLaunchTests`.
 - Tests for `LaunchListResponse`.
+- Tests for `LaunchResponse`.
 
 ### Changed